+++ log/helpers.py
@@ -2,20 +2,32 @@
 
 import logging
 
+from . filters import relpath_format_filter
+from . import state
+
 
 DEFAULT_LEVEL = logging.INFO
 DEFAULT_FORMAT = "%(levelname)s: %(name)s: %(message)s"
+VERBOSITY_TO_LEVEL = {
+    0: logging.ERROR,
+    1: logging.WARNING,
+    2: logging.INFO,
+    3: logging.DEBUG,
+}
 
-initialized = False
 
-
-def init(*, reset=False, debug=False, **kwargs):
+def init(*, reset=False, debug=False, verbosity=None, **kwargs):
     if reset:
         for handler in logging.root.handlers[:]:
             logging.root.removeHandler(handler)
 
     custom_format = kwargs.get('format')
-    default_level = logging.DEBUG if debug else DEFAULT_LEVEL
+    if debug:
+        default_level = logging.DEBUG
+    elif verbosity is not None:
+        default_level = VERBOSITY_TO_LEVEL[verbosity]
+    else:
+        default_level = DEFAULT_LEVEL
 
     kwargs['level'] = kwargs.get('level', default_level)
     kwargs['format'] = kwargs.get('format', DEFAULT_FORMAT)
@@ -26,8 +38,14 @@ def init(*, reset=False, debug=False, **kwargs):
         for handler in logging.root.handlers:
             handler.setFormatter(formatter)
 
-    global initialized
-    initialized = True
+    install_additional_formats(logging.root)
+
+    state.initialized = True
+
+
+def install_additional_formats(logger):
+    for handler in logger.handlers:
+        handler.addFilter(relpath_format_filter)
 
 
 def silence(*names, allow_info=False, allow_warning=False, allow_error=False):