+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision History
 
+## 1.1 (2018-10-26)
+
+- Added `%(relpath)s` logging format.
+- Added `verbosity` as `init()` option to work with Django admin commands.
+
 ## 1.0 (2018-09-27)
 
 - Initial stable release.